Feat: create profile context
